refactor: remove kubectl dependency in chaos test, use k8s API instead (#3501)
