+++ .github/workflows/chaos.yml
@@ -14,7 +14,7 @@ jobs:
 
       - name: Creating minikube cluster
         run: |
-          bash ./t/chaos/utils.sh start_minikube
+          bash ./t/chaos/setup_chaos_utils.sh start_minikube
 
       - name: Print cluster information
         run: |
@@ -29,17 +29,17 @@ jobs:
           git clone https://github.com/api7/etcd-operator.git --depth 1
           bash etcd-operator/example/rbac/create_role.sh
           kubectl create -f etcd-operator/example/deployment.yaml
-          bash ./t/chaos/utils.sh ensure_pods_ready etcd-operator "True" 30
+          bash ./t/chaos/setup_chaos_utils.sh ensure_pods_ready etcd-operator "True" 30
           kubectl create -f etcd-operator/example/example-etcd-cluster.yaml
-          bash ./t/chaos/utils.sh ensure_pods_ready etcd "True True True" 30
+          bash ./t/chaos/setup_chaos_utils.sh ensure_pods_ready etcd "True True True" 30
 
       - name: Deploy APISIX
         run: |
-          bash ./t/chaos/utils.sh modify_config
+          bash ./t/chaos/setup_chaos_utils.sh modify_config
           kubectl create configmap apisix-gw-config.yaml --from-file=./conf/config.yaml
           kubectl apply -f ./kubernetes/deployment.yaml
           kubectl apply -f ./kubernetes/service.yaml
-          bash ./t/chaos/utils.sh ensure_pods_ready apisix-gw "True" 30
+          bash ./t/chaos/setup_chaos_utils.sh ensure_pods_ready apisix-gw "True" 30
           nohup kubectl port-forward svc/apisix-gw-lb 9080:9080 >/dev/null 2>&1 &
 
       - name: Deploy Chaos mesh
